src/dom_tree/ops.rs: move traverse operations into `dom_tree/travesal.rs`
